Incomplete - task 33: Add a (x,y) (time,speed) and a (x,y) (time,distance) graph to the stats screen","Type-Enhancement, Priority-Medium, Milestone-Release0.9 http://code.google.com/p/open-gpstracker/issues/detail?id=33
Graphing time/speed over multiple segments

+++ application/src/nl/sogeti/android/gpstracker/viewer/LoggerMap.java
@@ -1070,7 +1070,7 @@ public class LoggerMap extends MapActivity
       for (int i = 0; i < mSpeedtexts.length; i++)
       {
          mSpeedtexts[i].setVisibility( View.VISIBLE );
-         int speed = (int) ( ( avgSpeed * 2d ) / 5d ) * i;
+         double speed = ( ( avgSpeed * 2d ) / 5d ) * i;
          String speedText = String.format( "%.0f %s", speed, mUnits.getSpeedUnit() ) ;
          mSpeedtexts[i].setText( speedText );
       }
